refactor(engine): move the #542 public/private boundary into gittensory-engine (#4883)
src/signals/redaction.ts is the canonical public/private text boundary (#542):
the PUBLIC_UNSAFE_TERMS / PUBLIC_LOCAL_PATH_INLINE vocabulary and isPublicSafeText,
which every public surface must run text through before surfacing it. It is
confirmed-pure, dependency-free string logic (no D1, no Env, no network) with the
highest fan-in of the not-yet-moved signals files -- 17 consumers across the
review stack -- so per #4883 it moves into the shared engine.

Relocate the vocabulary + guards into packages/gittensory-engine/src/signals/
redaction.ts verbatim and re-export them from the engine barrel. src/signals/
redaction.ts becomes a re-export shim, so all 17 existing `../signals/redaction`
import paths stay byte-for-byte unchanged. Behavior is identical; the existing
redaction suite now exercises the engine source through the shim.
